// apps/backend/src/modules/hardware-request/domain/enums/activity-action.enum.ts
export enum ActivityAction {
    CREATED = 'CREATED',
    UPDATED = 'UPDATED',
    SUBMITTED = 'SUBMITTED',
    REVIEWED = 'REVIEWED',
    APPROVED = 'APPROVED',
    REJECTED = 'REJECTED',
    CANCELLED = 'CANCELLED',
    PROCUREMENT_UPDATED = 'PROCUREMENT_UPDATED',
    PROCUREMENT_COMPLETED = 'PROCUREMENT_COMPLETED',
    INSTALL_SCHEDULED = 'INSTALL_SCHEDULED',
    INSTALL_CONFIRMED = 'INSTALL_CONFIRMED',
    INSTALL_RESCHEDULED = 'INSTALL_RESCHEDULED',
    INSTALL_STARTED = 'INSTALL_STARTED',
    INSTALL_COMPLETED = 'INSTALL_COMPLETED',
    COMMENTED = 'COMMENTED',
    BARCODE_SCANNED = 'BARCODE_SCANNED',
    CLOSED = 'CLOSED',
}